simplify the relationship between csrm and immutable service
